feat: Add support for alternative domains
This commit introduces a new feature that allows specifying alternative
domains to scrape from. This is useful for sites that host content,
such as images, on different domains.

A new `--alternative-domains` (or `-a`) flag has been added to specify
a comma-separated list of alternative domains.

The scraper will now process the main site first, and then the
alternative domains. The UI and save/resume functionality have also
been updated to support this new feature.